refactor(te): extract policy logic
extract the logic into its own module, leaving the genserver only for
message handling
